added sounds; boots behavior; level restart
