--- CS3540-Final-Game/Assets/PlayerAttack.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: e4d16516fbc79634393b25e397310b29
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
