+++ CS3540-Final-Game/Assets/Prefabs/BreakableRocks.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 548f74b02963f6a4ea6d00b356f758e7
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
