+++ CS3540-Final-Game/Assets/Prefabs/BreakableRocks/SM_Env_Rock_02 (1).prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 083a9191d488e7442ad3f4eb07d84941
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
