+++ CS3540-Final-Game/Assets/Prefabs/BreakableRocks/SM_Env_Rock_Huge_03 (1).prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: b31c8f45aecb6df4d9be810dbfd38e6c
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
