+++ CS3540-Final-Game/Assets/Prefabs/BreakableRocks/SM_Env_Rock_Large_03 (1).prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 09eb814636c74eb458ea4c607ad09596
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
