+++ CS3540-Final-Game/Assets/Prefabs/BreakableRocks/SM_Env_Rock_Large_05 (2).prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: e4798f018f180084ab137ac2a259866a
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
