+++ CS3540-Final-Game/Assets/Prefabs/BreakableRocks/SM_Env_Rock_Large_05 (3).prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 93d69dceefcf1b24492458b26c7744a8
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
