+++ CS3540-Final-Game/Assets/Prefabs/BreakableRocks/SM_Env_Rock_Large_06 (1).prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: df04e671a22f4f74f9f76c839016dde3
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
