+++ CS3540-Final-Game/Assets/Prefabs/BreakableRocks/SM_Env_Rock_Large_07 (1).prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 4f49a361640433144b804e5c4be98645
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
