+++ CS3540-Final-Game/Assets/Prefabs/BreakableRocks/SM_Env_Rock_Large_09 (1).prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: d87d2448478af59458ade5d5e53587cf
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
