+++ CS3540-Final-Game/Assets/Prefabs/boots.prefab
@@ -1,0 +1,145 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!1001 &6017567714541378898
+PrefabInstance:
+  m_ObjectHideFlags: 0
+  serializedVersion: 2
+  m_Modification:
+    serializedVersion: 3
+    m_TransformParent: {fileID: 0}
+    m_Modifications:
+    - target: {fileID: -8679921383154817045, guid: 30986e2f9fa62d6439d6a0718817a96a, type: 3}
+      propertyPath: m_LocalPosition.x
+      value: -21.16
+      objectReference: {fileID: 0}
+    - target: {fileID: -8679921383154817045, guid: 30986e2f9fa62d6439d6a0718817a96a, type: 3}
+      propertyPath: m_LocalPosition.y
+      value: 32.601
+      objectReference: {fileID: 0}
+    - target: {fileID: -8679921383154817045, guid: 30986e2f9fa62d6439d6a0718817a96a, type: 3}
+      propertyPath: m_LocalPosition.z
+      value: -84.139
+      objectReference: {fileID: 0}
+    - target: {fileID: -8679921383154817045, guid: 30986e2f9fa62d6439d6a0718817a96a, type: 3}
+      propertyPath: m_LocalRotation.w
+      value: 0.7071068
+      objectReference: {fileID: 0}
+    - target: {fileID: -8679921383154817045, guid: 30986e2f9fa62d6439d6a0718817a96a, type: 3}
+      propertyPath: m_LocalRotation.x
+      value: -0.7071068
+      objectReference: {fileID: 0}
+    - target: {fileID: -8679921383154817045, guid: 30986e2f9fa62d6439d6a0718817a96a, type: 3}
+      propertyPath: m_LocalRotation.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: -8679921383154817045, guid: 30986e2f9fa62d6439d6a0718817a96a, type: 3}
+      propertyPath: m_LocalRotation.z
+      value: -0
+      objectReference: {fileID: 0}
+    - target: {fileID: -8679921383154817045, guid: 30986e2f9fa62d6439d6a0718817a96a, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: -8679921383154817045, guid: 30986e2f9fa62d6439d6a0718817a96a, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: -8679921383154817045, guid: 30986e2f9fa62d6439d6a0718817a96a, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: -7511558181221131132, guid: 30986e2f9fa62d6439d6a0718817a96a, type: 3}
+      propertyPath: m_Materials.Array.data[0]
+      value: 
+      objectReference: {fileID: 2100000, guid: b974b5f2a40f62049bff286d64fa300a, type: 2}
+    - target: {fileID: 919132149155446097, guid: 30986e2f9fa62d6439d6a0718817a96a, type: 3}
+      propertyPath: m_Name
+      value: STC_mining_boot
+      objectReference: {fileID: 0}
+    - target: {fileID: 919132149155446097, guid: 30986e2f9fa62d6439d6a0718817a96a, type: 3}
+      propertyPath: m_TagString
+      value: BootsPickUp
+      objectReference: {fileID: 0}
+    m_RemovedComponents: []
+    m_RemovedGameObjects: []
+    m_AddedGameObjects: []
+    m_AddedComponents:
+    - targetCorrespondingSourceObject: {fileID: 919132149155446097, guid: 30986e2f9fa62d6439d6a0718817a96a, type: 3}
+      insertIndex: -1
+      addedObject: {fileID: 4103788151598684540}
+    - targetCorrespondingSourceObject: {fileID: 919132149155446097, guid: 30986e2f9fa62d6439d6a0718817a96a, type: 3}
+      insertIndex: -1
+      addedObject: {fileID: 7930244845927736093}
+    - targetCorrespondingSourceObject: {fileID: 919132149155446097, guid: 30986e2f9fa62d6439d6a0718817a96a, type: 3}
+      insertIndex: -1
+      addedObject: {fileID: 7754723914283159718}
+  m_SourcePrefab: {fileID: 100100000, guid: 30986e2f9fa62d6439d6a0718817a96a, type: 3}
+--- !u!1 &6864567913975914499 stripped
+GameObject:
+  m_CorrespondingSourceObject: {fileID: 919132149155446097, guid: 30986e2f9fa62d6439d6a0718817a96a, type: 3}
+  m_PrefabInstance: {fileID: 6017567714541378898}
+  m_PrefabAsset: {fileID: 0}
+--- !u!54 &4103788151598684540
+Rigidbody:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 6864567913975914499}
+  serializedVersion: 4
+  m_Mass: 1
+  m_Drag: 0
+  m_AngularDrag: 0.05
+  m_CenterOfMass: {x: 0, y: 0, z: 0}
+  m_InertiaTensor: {x: 1, y: 1, z: 1}
+  m_InertiaRotation: {x: 0, y: 0, z: 0, w: 1}
+  m_IncludeLayers:
+    serializedVersion: 2
+    m_Bits: 0
+  m_ExcludeLayers:
+    serializedVersion: 2
+    m_Bits: 0
+  m_ImplicitCom: 1
+  m_ImplicitTensor: 1
+  m_UseGravity: 1
+  m_IsKinematic: 0
+  m_Interpolate: 0
+  m_Constraints: 0
+  m_CollisionDetection: 0
+--- !u!65 &7930244845927736093
+BoxCollider:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 6864567913975914499}
+  m_Material: {fileID: 0}
+  m_IncludeLayers:
+    serializedVersion: 2
+    m_Bits: 0
+  m_ExcludeLayers:
+    serializedVersion: 2
+    m_Bits: 0
+  m_LayerOverridePriority: 0
+  m_IsTrigger: 1
+  m_ProvidesContacts: 0
+  m_Enabled: 1
+  serializedVersion: 3
+  m_Size: {x: 0.5264033, y: 0.30458003, z: 0.34594214}
+  m_Center: {x: 0.000000044703484, y: -0.064455286, z: 0.17310178}
+--- !u!114 &7754723914283159718
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 6864567913975914499}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: a689249b09e36f04fb07c418fbff4491, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  pickupSound: {fileID: 8300000, guid: 05dea694f7d94b7408d288d63a9b99ce, type: 3}
+  rotationAmount: 45
+  bobbingHeight: 0.1
+  bobbingSpeed: 1
