+++ CS3540-Final-Game/Assets/Prefabs/boots.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 6d435efcae981f24ca58c332562fce96
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
