--- CS3540-Final-Game/Assets/Prefabs/rock_barrier (1) 1.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 0ec92a7b833ce434a9526941e448848a
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
