--- CS3540-Final-Game/Assets/Prefabs/rock_barrier (1).prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 975a7dea653e79d46813a62a46618f92
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
