+++ CS3540-Final-Game/Assets/Scripts/BootsPickUpBehavior.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a689249b09e36f04fb07c418fbff4491
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
