+++ CS3540-Final-Game/Assets/Scripts/PlayerController.cs
@@ -9,6 +9,8 @@ public class PlayerController : MonoBehaviour
     public float walkSpeed = 5.0f;
     public float sprintSpeedScalar = 2.0f;
     public float jumpForceScalar = 2.0f;
+
+    public float superJumpForceScalar = 3f;
     public float gravity = 9.81f;
     public float airControl = 0.75f;
     CharacterController cc;
@@ -23,11 +25,12 @@ public class PlayerController : MonoBehaviour
     {
         cc = GetComponent<CharacterController>();
         m_Animator = gameObject.GetComponent<Animator>();
+        PlayerHealth.isDead = false;
 
     }
 
     // Update is called once per frame
-    void Update()
+    void FixedUpdate()
     {
         if (!PlayerHealth.isDead)
         {
@@ -37,8 +40,8 @@ public class PlayerController : MonoBehaviour
         if (transform.position.y < minHeight)
         {
             PlayerHealth.isDead = true;
-            // Would restart level in the future
-            Destroy(gameObject);
+            FindObjectOfType<LevelManager>().LevelLost();
+            Destroy(gameObject, 2f);
         }
     }
 
@@ -47,8 +50,12 @@ public class PlayerController : MonoBehaviour
     {
         float moveHorizontal = Input.GetAxis("Horizontal");
         float moveVertical = Input.GetAxis("Vertical");
+
         moveSpeed = Input.GetKey(KeyCode.LeftShift) && cc.isGrounded ? walkSpeed * sprintSpeedScalar : walkSpeed;
 
+        float jumpAmount = Input.GetKey(KeyCode.LeftShift) && cc.isGrounded && LevelManager.bootsPickedUp ? superJumpForceScalar : jumpForceScalar;
+
+        
         input = (transform.right * moveHorizontal + transform.forward * moveVertical).normalized;
         input *= moveSpeed;
 
@@ -82,7 +89,7 @@ public class PlayerController : MonoBehaviour
             if (Input.GetButton("Jump"))
             {
 
-                moveDirection.y = Mathf.Sqrt(2 * jumpForceScalar * gravity);
+                moveDirection.y = Mathf.Sqrt(2 * jumpAmount * gravity);
                 JumpAnimation();
             }
             else
