--- CS3540-Final-Game/Assets/WeaponAttack.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 825e639ae3c4adf47b5dc25338e7a15e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
